feat: Handle different types of public keys
Prior to this commit, only one type of verification key was supported:
ECDSA keys, ASN.1 DER-encoded, using the P-256 curve and SHA-256 as digest algorithm.

This is the key format used by default by `cosign` when generating a
pub/private key pair.

However, `cosign` can produce signatures using a user-provided private
key, which can be using a different algorithm.

This commit extends the range of supported keys to be:

  * RSA keys of 2048-8192 bits, PKCS#1.5 padding, and SHA-256 as digest algorithm
  * RSA keys of 2048-8192 bits, PKCS#1.5 padding, and SHA-384 as digest algorithm
  * RSA keys of 2048-8192 bits, PKCS#1.5 padding, and SHA-512 as digest algorithm
  * Ed25519 keys, and SHA-512 as the digest algorithm
  * ECDSA keys, ASN.1 DER-encoded, using the P-256 curve and SHA-256 as digest algorithm
  * ECDSA keys, ASN.1 DER-encoded, using the P-384 curve and SHA-384 as digest algorithm

This is achieved by leveraging the `ring` crate.

While working on this feature, the old `crpto.rs` has been split into
smaller, and more maintainable, files.
