+++ Cargo.toml
@@ -16,15 +16,15 @@ rustls-tls = ["oci-distribution/rustls-tls"]
 async-trait = "0.1.51"
 base64 = "0.13.0"
 der-parser = "6.0.1"
-ecdsa = { version = "0.12.4", features = ["verify", "pem", "der", "pkcs8"] }
 # TODO: go back to the officially release oci-distribution once these patches are released
-#oci-distribution = { version = "0.8.1", default-features = false }
 oci-distribution = { git = "https://github.com/krustlet/oci-distribution", rev = "0f717968093a5415f428503d741dedf24ea97948", default-features = false }
+#oci-distribution = { version = "0.8.1", default-features = false }
 olpc-cjson = "0.1.1"
-p256 = {version = "0.9.0", features = ["ecdsa-core"]}
-sha2 = "0.10.1"
+pem = "1.0.2"
+ring = "0.16.20"
 serde_json = "1.0.68"
 serde = {version = "1.0.130", features = ["derive"]}
+sha2 = "0.10.1"
 thiserror = "1.0.30"
 tokio = { version = "1.12.0", features = ["full"]}
 tough = { version = "0.12.1", features = [ "http" ] }
