+++ src/cosign/bundle.rs
@@ -17,7 +17,7 @@ use olpc_cjson::CanonicalFormatter;
 use serde::{Deserialize, Serialize};
 use std::cmp::PartialEq;
 
-use crate::crypto::{verify_signature, CosignVerificationKey};
+use crate::crypto::{CosignVerificationKey, Signature};
 use crate::errors::{Result, SigstoreError};
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
@@ -46,7 +46,10 @@ impl Bundle {
             ))
         })?;
 
-        verify_signature(rekor_pub_key, &bundle.signed_entry_timestamp, &buf)?;
+        rekor_pub_key.verify_signature(
+            Signature::Base64Encoded(bundle.signed_entry_timestamp.as_bytes()),
+            &buf,
+        )?;
         Ok(bundle)
     }
 }
@@ -67,7 +70,7 @@ mod tests {
     use serde_json::json;
 
     use crate::cosign::tests::get_rekor_public_key;
-    use crate::crypto;
+    use crate::crypto::SignatureDigestAlgorithm;
 
     fn build_correct_bundle() -> String {
         let bundle_json = json!({
@@ -98,7 +101,11 @@ mod tests {
 MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAENptdY/l3nB0yqkXLBWkZWQwo6+cu
 OSWS1X9vPavpiQOoTTGC0xX57OojUadxF1cdQmrsiReWg2Wn4FneJfa8xw==
 -----END PUBLIC KEY-----"#;
-        let not_rekor_pub_key = crypto::new_verification_key(public_key).unwrap();
+        let not_rekor_pub_key = CosignVerificationKey::from_pem(
+            public_key.as_bytes(),
+            SignatureDigestAlgorithm::default(),
+        )
+        .expect("Cannot create CosignVerificationKey");
 
         let bundle_json = build_correct_bundle();
         let bundle = Bundle::new_verified(&bundle_json, &not_rekor_pub_key);
